+++ src/components/Breadcrumb/index.vue
@@ -1,9 +1,11 @@
 <template>
-  <el-breadcrumb class="app-levelbar" separator="/">
+  <el-breadcrumb class="app-breadcrumb" separator="/">
+  <transition-group name="breadcrumb">
     <el-breadcrumb-item v-for="(item,index)  in levelList" :key="item.path">
       <router-link v-if='item.redirect==="noredirect"||index==levelList.length-1' to="" class="no-redirect">{{item.name}}</router-link>
       <router-link v-else :to="item.redirect||item.path">{{item.name}}</router-link>
     </el-breadcrumb-item>
+     </transition-group>
   </el-breadcrumb>
 </template>
 
@@ -36,7 +38,7 @@ export default {
 </script>
 
 <style rel="stylesheet/scss" lang="scss" scoped>
-.app-levelbar.el-breadcrumb {
+.app-breadcrumb.el-breadcrumb {
   display: inline-block;
   font-size: 14px;
   line-height: 50px;
