+++ src/views/layout/components/Sidebar/SidebarItem.vue
