+++ src/views/layout/components/index.js
@@ -1,7 +1,3 @@
 export { default as Navbar } from './Navbar'
-
 export { default as Sidebar } from './Sidebar'
-
-export { default as Levelbar } from './Levelbar'
-
 export { default as AppMain } from './AppMain'
